Extract infinity support into django_pg_infinity package
