+++ src/argus/incident/fields.py
@@ -1,46 +1,11 @@
-from datetime import datetime
-
 from django import forms
-from django.db import models
 from django.forms import fields
 
 import rest_framework.fields as rest_framework_fields
 
-import argus.util.datetime_utils as utils
+import django_pg_infinity.utils as utils
+from django_pg_infinity.fields import DateTimeInfinityField  # noqa: F401 — re-exported for backwards compatibility
 from .validators import validate_key_value
-
-
-class DateTimeInfinityField(models.DateTimeField):
-    # Code based on https://github.com/Uninett/nav/blob/44a67a5037305c946eb69666d0a4b3b51ea5cff4/python/nav/models/fields.py#L41-L53
-    def get_db_prep_value(self, value, connection, prepared=False):
-        is_postgres = self.is_postgres(connection)
-
-        if isinstance(value, datetime):
-            infinity_repr = utils.get_infinity_repr(value, str_repr=is_postgres)
-            if infinity_repr:
-                # (Presumably) only PostgreSQL accepts - and correctly adapts - infinity strings
-                if is_postgres:
-                    return connection.ops.adapt_datetimefield_value(infinity_repr)
-                return infinity_repr
-        elif isinstance(value, str):
-            infinity_time = utils.parse_infinity(value, return_localized=False)
-            if infinity_time:
-                return value if is_postgres else infinity_time
-
-        return super().get_db_prep_value(value, connection, prepared=prepared)
-
-    def from_db_value(self, value, expression, connection):
-        if value is None:
-            return value
-
-        return utils.get_infinity_time(value) or utils.convert_datetimefield_value(value, connection)
-
-    def to_python(self, value):
-        return utils.get_infinity_time(value) or super().to_python(value)
-
-    @staticmethod
-    def is_postgres(connection):
-        return connection.vendor == "postgresql"
 
 
 class SplitDateTimeInfinityField(forms.SplitDateTimeField):
